Added functionality for finding prime numbers.

+++ ModernJava/src/main/java/com/github/onelineoff/fp/util/Utils.java
@@ -2,5 +2,5 @@ package com.github.onelineoff.fp.util;
 
 public interface Utils {
 	
-	public boolean isPrime(Integer val);
+	public boolean isPrime(Long val);
 }
